pyproject.toml
[project]
name = "flock-git"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flock-core>=0.5.30",
]
